Make the focused window always visible on the screen when moving it to a new created workspace
Before this commit, using the "Move Window to New Workspace" shortcut created a new workspace and teleported the active window into it. When GNOME Shell started playing the workspace switch animation, the active window disappeared from view. Then, the workspace switched to the new workspace to the window that had just been moved. Using the shortcut makes it look like the active window disappears and reappears on the screen.

GNOME Shell already has a built-in shortcut that moves a window to a workspace, and the window remains visible on the screen during the workspace switch animation. This commit references GNOME Shell's official behavior to make this extension's window switching animation feel the same.

Instead of teleporting the focused/active window the a new created workspace it simply pins the window to all workspaces which it makes always visible on the screen when switching to other workspaces. Visually it looks the same as the official GNOME Shell's window moving to a workspace animation.